[ cleanup ] move coverage to separate module
